+++ tests/conftest.py
@@ -11,6 +11,7 @@ import pytest
 from tests.test_data import (
     create_aggregate_over_rows_test_df,
     create_bool_df,
+    create_categorical_df,
     create_is_between_dates_df_1,
     create_numeric_df_1,
     create_numeric_df_2,
@@ -147,6 +148,10 @@ def minimal_attribute_dict():
         "_BooleanImputer": {
             "columns": ["b"],
             "impute_value": True,
+        },
+        "CategoricalImputer": {
+            "columns": ["b"],
+            "impute_value": "bla",
         },
         "CappingTransformer": {
             "capping_values": {"a": [0.1, 0.2]},
@@ -301,7 +306,7 @@ def minimal_attribute_dict():
             "new_column_name": "c",
             "separator": "-",
         },
-        "_StringImputer": {
+        "StringImputer": {
             "columns": ["b"],
             "impute_value": "missing",
         },
@@ -371,6 +376,7 @@ def minimal_dataframe_lookup(request) -> dict[str, pd.DataFrame]:
     num_df = create_numeric_df_1(library=library)
     nan_df = create_numeric_df_2(library=library)
     object_df = create_object_df(library=library)
+    categorical_df = create_categorical_df(library=library)
     date_df = create_is_between_dates_df_1(library=library)
     agg_df = create_aggregate_over_rows_test_df(
         library=library,
@@ -417,7 +423,8 @@ def minimal_dataframe_lookup(request) -> dict[str, pd.DataFrame]:
     min_df_dict["BaseAggregationTransformer"] = agg_df
     min_df_dict["AggregateRowsOverColumnTransformer"] = agg_df
     min_df_dict["WhenThenOtherwiseTransformer"] = when_then_df
-    min_df_dict["_StringImputer"] = object_df
+    min_df_dict["StringImputer"] = object_df
+    min_df_dict["CategoricalImputer"] = categorical_df
     min_df_dict["_BooleanImputer"] = bool_df
 
     pyarrow_transformers = ["ExtractStringComponentsTransformer"]